requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altgraph==0.17.4
asgiref==3.5.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
Django==4.0.5
djangorestframework==3.13.1
Flask==3.1.0
frozenlist==1.6.0
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
ngrok==1.4.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
propcache==0.3.1
pyee==11.1.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2022.1
pywin32-ctypes==0.2.3
requests==2.28.0
six==1.17.0
slack_sdk==3.35.0
slackeventsapi==3.0.3
sqlparse==0.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.9
Werkzeug==3.1.3
yarl==1.20.0
